fix: linking edit category
